:bug: 判断是否有 team

+++ nonebot_plugin_osubot/draw/utils.py
@@ -267,18 +267,18 @@ async def update_icon(info: UnifiedUser):
                 user_icon = await get_projectimg(info.avatar_url)
                 with open(path / f"icon.{info.avatar_url.split('.')[-1]}", "wb") as f:
                     f.write(user_icon.getvalue())
-
-    team_path = team_cache_path / f"{info.team.id}.png"
-    if team_path.exists():
-        modified_time = team_path.stat().st_mtime
-        modified_datetime = datetime.datetime.fromtimestamp(modified_time)
-        time_diff = datetime.datetime.now() - modified_datetime
-        # 判断文件是否创建超过一天
-        if time_diff > datetime.timedelta(days=1):
-            team_path.unlink()
-            team_icon = await get_projectimg(info.team.flag_url)
-            with open(team_path, "wb") as f:
-                f.write(team_icon.getvalue())
+    if info.team:
+        team_path = team_cache_path / f"{info.team.id}.png"
+        if team_path.exists():
+            modified_time = team_path.stat().st_mtime
+            modified_datetime = datetime.datetime.fromtimestamp(modified_time)
+            time_diff = datetime.datetime.now() - modified_datetime
+            # 判断文件是否创建超过一天
+            if time_diff > datetime.timedelta(days=1):
+                team_path.unlink()
+                team_icon = await get_projectimg(info.team.flag_url)
+                with open(team_path, "wb") as f:
+                    f.write(team_icon.getvalue())
 
 
 async def update_map(set_id, map_id):
